chore: Renamed Core/RichClientRequest.mts to Core/rich-client-request.mts.
