+++ packages/architectura/src/Core/rich-client-request.mts
